Squashed 'external/toxcore/c-toxcore/' changes from 03e9fbf3703..55752a2e2ef
55752a2e2ef fix(toxav): pass video bit rate as kbit Previously we unintentionally made it Mbit.
7e573280a75 docs(toxav): fix docs of toxav.h - fix units to be more readable - use width before height consistently - video -> audio typo
5f88a084e8c fix: friend_connections leak on allocation failure clean up when it only contains connections in the NONE state
6d27a1ae178 fix: wrong comment for closelist
ce4f29e8036 cleanup: Fix all `-Wsign-compare` warnings.
4d4251c397f chore: lower cirrus ci timeout drastically
40676284507 fix: events leak that can occur if allocation fails rare in practice, found by fuzzing
9610ac31c5f fix: Return an error instead of crashing on nullptr args in NGC.
a57c2c8f956 refactor: Make ToxAV independent of toxcore internals.
5752fc29f86 refactor: Make tox-bootstrapd use bool instead of int
df675786eb2 chore: Add release-drafter github action.
03fd7a69dcf chore: Use toktok's cmp instead of upstream.
350c0ba1205 cleanup: Sort apk/apt install commands in Dockerfiles.
8c1bda502cb chore(deps): bump golang.org/x/net
ddb9d3210da chore: Upgrade to FreeBSD 14.1 in cirrus build.
e9076f45bd3 chore(cmake): set options changes as cache and with force

git-subtree-dir: external/toxcore/c-toxcore
git-subtree-split: 55752a2e2ef894bfa6d7a2a21a0278e3f2bede7d

+++ other/fun/sign.c
@@ -35,7 +35,7 @@ static int load_file(const char *filename, unsigned char **result)
     fseek(f, 0, SEEK_SET);
     *result = (unsigned char *)malloc(size + 1);
 
-    if (size != fread(*result, sizeof(char), size, f)) {
+    if ((size_t)size != fread(*result, sizeof(char), size, f)) {
         free(*result);
         fclose(f);
         return -2; // -2 means file reading fail
@@ -55,13 +55,13 @@ int main(int argc, char *argv[])
         crypto_sign_ed25519_keypair(pk, sk);
         printf("Public key:\n");
 
-        for (int i = 0; i < crypto_sign_ed25519_PUBLICKEYBYTES; ++i) {
+        for (uint32_t i = 0; i < crypto_sign_ed25519_PUBLICKEYBYTES; ++i) {
             printf("%02X", pk[i]);
         }
 
         printf("\nSecret key:\n");
 
-        for (int i = 0; i < crypto_sign_ed25519_SECRETKEYBYTES; ++i) {
+        for (uint32_t i = 0; i < crypto_sign_ed25519_SECRETKEYBYTES; ++i) {
             printf("%02X", sk[i]);
         }
